139064652 add debug and trace levels
Add debug and trace levels to the central logger for both system and event logs.
Alter the order of if check for better performance.
trace levels can be used for printing error stacks to log file.
debug level can be used to print any variable values in log file.
The new order of logging is:
trace
debug
info
warn
error